Refinery/ByTrying-Transform: ByTrying is Transformation, not constraint, move from logical to root
