+++ src/Refinery/Custom/Constraint.php
@@ -8,11 +8,13 @@ use ILIAS\Refinery\DeriveTransformFromApplyTo;
 use ILIAS\Refinery\DeriveInvokeFromTransform;
 use ILIAS\Data;
 use ILIAS\Data\Result;
+use ILIAS\Refinery\ProblemBuilder;
 
 class Constraint implements ConstraintInterface
 {
     use DeriveTransformFromApplyTo;
     use DeriveInvokeFromTransform;
+    use ProblemBuilder;
 
     /**
      * @var ILIAS\Data\Factory
@@ -107,63 +109,5 @@ class Constraint implements ConstraintInterface
         }
 
         return $result;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    final public function withProblemBuilder(callable $builder)
-    {
-        $clone = clone $this;
-        $clone->error = $builder;
-        return $clone;
-    }
-
-    /**
-     * Get the problem message
-     *
-     * @return string
-     */
-    final public function getErrorMessage($value)
-    {
-        $lng_closure = $this->getLngClosure();
-        return call_user_func($this->error, $lng_closure, $value);
-    }
-
-    /**
-     * Get the closure to be passed to the error-function that does i18n and
-     * sprintf.
-     *
-     * @return	\Closure
-     */
-    final protected function getLngClosure()
-    {
-        return function () {
-            $args = func_get_args();
-            if (count($args) < 1) {
-                throw new \InvalidArgumentException(
-                    "Expected an id of a lang var as first parameter"
-                );
-            }
-            $error = $this->lng->txt($args[0]);
-            if (count($args) > 1) {
-                $args[0] = $error;
-                for ($i = 0; $i < count($args); $i++) {
-                    $v = $args[$i];
-                    if ((is_array($v) || is_object($v) || is_null($v))
-                    && !method_exists($v, "__toString")) {
-                        if (is_array($v)) {
-                            $args[$i] = "array";
-                        } elseif (is_null($v)) {
-                            $args[$i] = "null";
-                        } else {
-                            $args[$i] = get_class($v);
-                        }
-                    }
-                }
-                $error = call_user_func_array("sprintf", $args);
-            }
-            return $error;
-        };
     }
 }
